requirements.txt
tensorflow-gpu
numpy
opencv-python
tqdm
pillow
pypiwin32